requirements.txt
google-cloud-pubsub
functions-framework
google-cloud-logging

